requirements.txt
google-generativeai
bce-embedding
faiss-cpu
Pillow
numpy
gradio
tqdm
python-dotenv
